pyproject.toml
[project]
name = "quantigence"
version = "0.1.0"
description = ""
authors = [
    {name = "alpha-coder38",email = "paymodeharshad00@gmail.com"}
]
readme = "README.md"



requires-python = ">=3.11, <3.13"
dependencies = [
    "redis (>=7.1.0,<8.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-core (>=1.1.0,<2.0.0)",
    "llama-index (>=0.14.8,<0.15.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "edgartools (>=4.35.0,<5.0.0)",
    "frozendict (>=2.4.7,<3.0.0)",
    "path (>=17.1.1,<18.0.0)",
    "cssutils (>=2.11.1,<3.0.0)",
    "neo4j (==5.28.2)",
    "langchain-neo4j (==0.6.0)",
    "neo4j-graphrag[mistralai] (>=1.11.0,<2.0.0)",
    "reportlab (==4.4.3)",
    "mistralai (>=1.10.0,<2.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "mne (>=1.11.0,<2.0.0)",
    "azure-core (>=1.38.0,<2.0.0)",
    "azure-ai-inference (>=1.0.0b9,<2.0.0)",
    "langgraph (>=1.0.6,<2.0.0)",
    "langsmith (>=0.6.4,<0.7.0)",
    "langchain-openai (>=1.1.7,<2.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "statsmodels (>=0.14.6,<0.15.0)",
    "yfinance (>=1.0,<2.0)",
    "financetoolkit (>=2.0.6,<3.0.0)",
    "nemoguardrails[all] (>=0.20.0,<0.21.0)",
    "datasets (>=4.5.0,<5.0.0)",
    "torch (>=2.10.0,<3.0.0)",
    "llmlingua (>=0.2.2,<0.3.0)",
    "transformers (==4.35.2)",
    "sentence-transformers (==3.0.0)",
    "openai (>=2.17.0,<3.0.0)",
    "deepeval (==3.4.5)",
    "azure-cosmos (>=4.14.6,<5.0.0)",
    "fastapi[standard] (>=0.129.0,<0.130.0)",
    "redis-om (>=1.0.6,<2.0.0)",
    "aredis (>=1.1.8,<2.0.0)",
    "azure-storage-blob (>=12.28.0,<13.0.0)",
    "azure-mgmt-storage (>=24.0.0,<25.0.0)",
    "azure-identity (>=1.25.2,<2.0.0)",
    "opencensus-ext-azure (>=1.1.15,<2.0.0)",
]

[tool.poetry]
packages = [{include = "quantigence", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ipykernel (>=7.1.0,<8.0.0)"
]
